+++ src/Upload.jsx
@@ -2,10 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import AjaxUpload from './AjaxUploader';
 import IframeUpload from './IframeUploader';
-let FlashUploader = null
-if(window.SWFUpload){
-  FlashUploader = require('./FlashUploader');
-}
+import FlashUploader from './FlashUploader';
 
 function empty() {
 }
@@ -39,6 +36,11 @@ class Upload extends Component {
     onReady: PropTypes.func,
     withCredentials: PropTypes.bool,
     supportServerRender: PropTypes.bool,
+    openFileDialogOnClick: PropTypes.bool,
+    flash: PropTypes.oneOfType([
+      PropTypes.object,
+      PropTypes.bool,
+    ]),
   }
 
   static defaultProps = {
@@ -57,6 +59,8 @@ class Upload extends Component {
     beforeUpload: null,
     customRequest: null,
     withCredentials: false,
+    openFileDialogOnClick: true,
+    flash: true,
   }
 
   state = {
@@ -73,23 +77,36 @@ class Upload extends Component {
   }
 
   getComponent() {
-    return typeof File !== 'undefined' ? AjaxUpload : (FlashUploader && 'flash' in this.props ? FlashUploader : IframeUpload);
+    let FinalUploader = null;
+    if (typeof File !== 'undefined') {
+      FinalUploader = AjaxUpload;
+    } else if (FlashUploader && this.props.flash) {
+      FinalUploader = FlashUploader;
+    } else {
+      FinalUploader = IframeUpload;
+    }
+
+    return FinalUploader;
   }
 
   abort(file) {
-    this.refs.inner.abort(file);
+    this.uploader.abort(file);
+  }
+
+  saveUploader = (node) => {
+    this.uploader = node;
   }
 
   render() {
     if (this.props.supportServerRender) {
       const ComponentUploader = this.state.Component;
       if (ComponentUploader) {
-        return <ComponentUploader {...this.props} ref="inner"/>;
+        return <ComponentUploader {...this.props} ref={this.saveUploader} />;
       }
       return null;
     }
     const ComponentUploader = this.getComponent();
-    return <ComponentUploader {...this.props} directory={this.props.directory} ref="inner"/>;
+    return <ComponentUploader {...this.props} ref={this.saveUploader} />;
   }
 }